Added driver, optional subtable clauses
It's now possible to generate a random number of values from a subtable. To enable this, every subtable call needs braces specifying which part of the statement is "associated" with it and should be removed if it is not instantiated. (A simple default is to just wrap the subtable call itself.)

Also added a driver with a REPL loop to generate objects, so that I can actually use it easily during a game.

Plus some more file changes and support code for the files.