State Machine Probably Works
Finished up the state machine, I think. It's got a fair number of test cases, but I'm sure I'll find things that I've missed - it's late, and I'm not up to thinking up every edge case right now. Plus, state machines (especially with the janky nesting solution I'm using) have a LOT of edge cases.

I'd still like to find  a better solution for handing how to identify which part of the string maps to which part of the automata in order to parse it, but I want to take a stab at figuring out a solution myself before I look up the "right" way to do it.

Next steps: Writing the code to instantiate automata from the ABNF file, either extending the automata to a push down automata or finding another way to handle the non-context free parts of the grammar, and converting from the automata result into a parse tree. Hopefully once that's done the parsing will be more robust.