Parser control tokens and automata character classes
The parser now extracts character for optionality ('[' and ']'), repetition ('*', possibly prefixed and/or suffixed with numbers), and alternatives ('/'). This should pave the way for constructing the automata from the grammar file.

The automata now allows for character sets to be specified, in the form of ALPHA, DIGIT, or CHAR, to avoid having to branch off for every printable character every time we process a quoted string.