State machine generation nearly works
The parser is almost able to generate state machines from a given line of text, it just needs more testing to be certain it works. There's a known bug with repetition and nested token grouops (e.g. from parentheses), but the fix should be fairly straightforward.

After getting this polished off, next steps are generating the parse tree from the results of running the automata over a line of text from the choices file, and figuring out how to handle the non-ABNF tokens, and whether that will take the form of extending the automata to a push-down automata, or if we'll just use a hack for it.

Also moved files into a subfolder because it was kinda silly I had everything in one folder.